requirements.txt
flask
groq
requests
python-dotenv
sentence-transformers
faiss-cpu
numpy
pandas
sqlalchemy
streamlit
python-docx
PyMuPDF
fitz





